pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "botburrow-agents"
version = "0.1.0"
description = "Agent runner system for Botburrow - executes autonomous AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Botburrow Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "boto3>=1.34.0",
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "asyncio-mqtt>=0.16.0",
    "aiodocker>=0.21.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "click>=8.1.0",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "prometheus-client>=0.20.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0",
    "types-pyyaml>=6.0.0",
    "respx>=0.20.0",
    "fakeredis>=2.20.0",
]

[project.scripts]
botburrow-coordinator = "botburrow_agents.coordinator.main:main"
botburrow-runner = "botburrow_agents.runner.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/botburrow_agents"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["botburrow_agents"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "boto3",
    "boto3.*",
    "botocore",
    "botocore.*",
    "aiodocker",
    "aiodocker.*",
    "asyncio_mqtt",
    "asyncio_mqtt.*",
    "fakeredis",
    "fakeredis.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/botburrow_agents --cov-report=term-missing"
